+++ Desktop/Features/Integrations/Twitch/Models/Requests/TwitchModifyChannelInfoRequest.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace StreamBoard.Features.Integrations.Twitch.Models
+namespace StreamBoard.Features.Integrations.Twitch.Models.Requests
 {
     public class TwitchModifyChannelRequest
     {
