+++ Desktop/Features/Integrations/Twitch/Services/ApiModules/TwitchApiModerationModule.cs
@@ -1,8 +1,10 @@
 using System.Net.Http;
 using System.Net.Http.Json;
 using StreamBoard.Features.Integrations.Twitch.Models;
+using StreamBoard.Features.Integrations.Twitch.Models.Requests;
+using StreamBoard.Features.Integrations.Twitch.Models.Responses;
 
-namespace StreamBoard.Features.Integrations.Twitch.Services
+namespace StreamBoard.Features.Integrations.Twitch.Services.ApiModules
 {
     public class TwitchApiModerationModule(TwitchAuthContext context, HttpClient http)
         : TwitchApiModule(context, http)
