+++ Desktop/Features/Integrations/Twitch/Services/ApiModules/TwitchApiUsersModule.cs
@@ -2,8 +2,9 @@ using System.Net.Http;
 using System.Net.Http.Json;
 using StreamBoard.Features.Integrations.Twitch.Exceptions;
 using StreamBoard.Features.Integrations.Twitch.Models;
+using StreamBoard.Features.Integrations.Twitch.Models.Responses;
 
-namespace StreamBoard.Features.Integrations.Twitch.Services
+namespace StreamBoard.Features.Integrations.Twitch.Services.ApiModules
 {
     public class TwitchApiUsersModule(TwitchAuthContext context, HttpClient http)
         : TwitchApiModule(context, http)
